backend/requirements.txt
Flask==3.1.0
huggingface_hub==0.29.3
imgaug==0.4.0
matplotlib==3.10.1
numpy==2.2.5
Pillow==11.2.1
torch==2.6.0
